# cmd/complain/doc/swagger.yaml
definitions:
  model.RequestCreateComplain:
    properties:
      createDateAt:
        type: integer
      exhibitionID:
        type: string
      message:
        type: string
      status:
        $ref: '#/definitions/model.StatusEnum'
      user:
        $ref: '#/definitions/model.User'
    required:
    - exhibitionID
    - message
    type: object
  model.StatusEnum:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - Pending
    - Approved
    - Rejected
  model.User:
    properties:
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      profile:
        type: string
    type: object
info:
  contact: {}
  description: Complain Service สำหรับขอจัดการเกี่ยวกับ Workshop Manager ทั้งการสร้าง
    แก้ไข ลบ Workshop Manager
  title: Complain Service API
  version: v0
paths:
  /api-complains/complains:
    post:
      consumes:
      - application/json
      description: Create a new complain
      operationId: CreateComplain
      parameters:
      - description: Complain data to create
        in: body
        name: requestExhibition
        required: true
        schema:
          $ref: '#/definitions/model.RequestCreateComplain'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a new complain
      tags:
      - CreateComplains
  /api-complains/complains/all:
    get:
      consumes:
      - application/json
      description: GetAllComplains
      operationId: GetAllComplains
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: GetAllComplains
      tags:
      - GetComplains
  /api-complains/complains/exhibitions:
    get:
      consumes:
      - application/json
      description: GetComplainsGroupByExhibitionName
      operationId: GetComplainsGroupByExhibitionName
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: GetComplainsGroupByExhibitionName
      tags:
      - GetComplains
  /api-complains/complains/exhibitions/{id}:
    get:
      consumes:
      - application/json
      description: GetComplainByExhibitionID
      operationId: GetComplainByExhibitionID
      parameters:
      - description: Exhibition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: GetComplainByExhibitionID
      tags:
      - GetComplains
  /api-complains/complains/exhibitions/{id}/reject:
    delete:
      consumes:
      - application/json
      description: DeleteComplainByExhibitionID
      operationId: DeleteComplainByExhibitionID
      parameters:
      - description: Exhibition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: DeleteComplainByExhibitionID
      tags:
      - Rejects
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"